move `lowering` module into `rust_ir`
