+++ src/bin/chalki.rs
@@ -16,7 +16,7 @@ use std::process::exit;
 
 use chalk::ir;
 use chalk::rust_ir;
-use chalk::ir::lowering::*;
+use chalk::rust_ir::lowering::*;
 use chalk::solve::SolverChoice;
 use docopt::Docopt;
 use rustyline::error::ReadlineError;
