+++ src/ir.rs
@@ -13,7 +13,6 @@ mod macros;
 
 crate mod could_match;
 crate mod debug;
-pub mod lowering;
 pub mod tls;
 
 crate type Identifier = InternedString;
@@ -275,25 +274,6 @@ impl<T, L> ast::Kinded for ParameterKind<T, L> {
         match *self {
             ParameterKind::Ty(_) => ast::Kind::Ty,
             ParameterKind::Lifetime(_) => ast::Kind::Lifetime,
-        }
-    }
-}
-
-pub trait ToParameter {
-    /// Utility for converting a list of all the binders into scope
-    /// into references to those binders. Simply pair the binders with
-    /// the indices, and invoke `to_parameter()` on the `(binder,
-    /// index)` pair. The result will be a reference to a bound
-    /// variable of appropriate kind at the corresponding index.
-    fn to_parameter(&self) -> Parameter;
-}
-
-impl<'a> ToParameter for (&'a ParameterKind<()>, usize) {
-    fn to_parameter(&self) -> Parameter {
-        let &(binder, index) = self;
-        match *binder {
-            ParameterKind::Lifetime(_) => ParameterKind::Lifetime(Lifetime::Var(index)),
-            ParameterKind::Ty(_) => ParameterKind::Ty(Ty::Var(index)),
         }
     }
 }
