+++ src/rust_ir.rs
@@ -11,6 +11,8 @@ use ir::{
 use std::collections::BTreeMap;
 use std::iter;
 
+pub mod lowering;
+
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Program {
     /// From type-name to item-id. Used during lowering only.
